Added rotary encoder
A rotary encoder can now be used to quickly emit events depending on
rotation direction and current layer.

The encoder is connected to timer 1 pins pa8 and pa9, and uses the
quadriature decoding that is already available in the chip.

Adjusted command.c to allow it to be set later on + adjusted flash
load/save to be able to store the new settings.