Adjusted sk6812 footprint, started routing
Adjusted sk6812 footprint from square cutout to something that is
routable with a mill.

The kailh sockets and leds eat up all the space between the keys. This
means that there is not room for the micro, regulator and usb
receptacle if we stay in a 10x10.

Options:
- Increase size to 11x11 or even 12x12. This would allow for some nice
  mounting holes.

- Make another board beneath the switch board that contains the micro
  and friends.

Lets think about this some more.