Now scan the matrix per row
We now read the column input at the start of the matrix_scan function,
and enable the next row at the end of this function. This allows the
row signal to stabilise in between calls, and is more robust in the
face of bad board layout.

The function is more complex, as it can only build up an accurate
picture of the keyboard after 5 consecutive calls. Update rate is
still 20kHz, so "good enough for the girls I go out with".