Added rgb actions for key events
This includes setting a group of keys to a color if a key is
pressed. My usecase: I have a macropad that selects between 9
workspaces. I want the workspace keys to have the color of the
selected workspace. This way I have a visual reference of what
workspace is selected at a glance.