Drop ease tied to a key, goto ease tied to an event
I'd like to see the active desktop as a color under my 5x5x2. Make a
way for my window manager to tell the keyboard where we are at.

As a consequence I'm removing the ability to set colors to
keypresses. I think that is too simple, we want colors on events. And
while events can be keypresses, it can also be e.g. the host telling
us what just happened.