Add a slow rainbow as a background effect
The default one color only light as backlight is not as pretty as
unicorn colors moving slowly across the keyboard.

This is a first stab; next step could be to allow the user to say
background effect = fixed color, slow rainbow, pulsing color.

Let's try to life with this for a few days first.